+++ CHANGES.md
@@ -1,9 +1,11 @@
 **Added:**
-* project export notification
-* project save notification
+* none
 
 **Changed:**
-* none
+* removed "display type" from export dialog
+* made macro edit popup have grid pane layout
+* added labels to sound value editor
+* boolean editing in macro edit popup is centered rather than on left side
 
 **Fixed:**
 * none